Add icon for Add To Shapes Lab option in context menu

+++ PowerPointLabs/PowerPointLabs/ThisAddIn.cs
@@ -309,16 +309,24 @@ namespace PowerPointLabs
 
             if (pres.Saved == Office.MsoTriState.msoTrue)
             {
-                // remove task pane
+                if (_documentHashcodeMapper.ContainsKey(pres.Application.ActiveWindow))
+                {
+                    _documentHashcodeMapper.Remove(pres.Application.ActiveWindow);
+                }
+
+                // if there exists some task panes, remove them
+                if (!_documentPaneMapper.ContainsKey(pres.Application.ActiveWindow))
+                {
+                    return;
+                }
+
                 var activePanes = _documentPaneMapper[pres.Application.ActiveWindow];
                 foreach (var pane in activePanes)
                 {
                     CustomTaskPanes.Remove(pane);
                 }
 
-                // remove entry from mappers
                 _documentPaneMapper.Remove(pres.Application.ActiveWindow);
-                _documentHashcodeMapper.Remove(pres.Application.ActiveWindow);
             }
             else
             {
@@ -434,7 +442,7 @@ namespace PowerPointLabs
 
             _documentHashcodeMapper[activeWindow] = tempName;
 
-            RegisterTaskPane(new RecorderTaskPane(tempName), "Record Management", activeWindow,
+            RegisterTaskPane(new RecorderTaskPane(tempName), TextCollection.RecManagementPanelTitle, activeWindow,
                              TaskPaneVisibleValueChangedEventHandler, null);
         }
 
@@ -448,7 +456,7 @@ namespace PowerPointLabs
             var activeWindow = presentation.Application.ActiveWindow;
 
             TaskPaneSetup(presentation);
-            RegisterTaskPane(new ColorPane(), "Color Panel", activeWindow, null, null);
+            RegisterTaskPane(new ColorPane(), TextCollection.ColorsLabTaskPanelTitle, activeWindow, null, null);
         }
 
         public void RegisterCustomShapePane(PowerPoint.Presentation presentation)
@@ -463,7 +471,8 @@ namespace PowerPointLabs
             TaskPaneSetup(presentation);
             RegisterTaskPane(
                 new CustomShapePane(_defaultShapeMasterFolderPrefix + DefaultShapeMasterFolderName,
-                                    DefaultShapeCategoryName), "Shapes Lab", activeWindow, null, null);
+                                    DefaultShapeCategoryName), TextCollection.ShapesLabTaskPanelTitle, activeWindow,
+                                    null, null);
         }
 
         public void SyncShapeAdd(string shapeName, string shapeFullName)
@@ -886,8 +895,8 @@ namespace PowerPointLabs
         # region Copy paste handlers
 
         private PowerPoint.DocumentWindow _copyFromWnd;
-        private Regex _shapeNamePattern = new Regex(@"^[^\[]\D+\s\d+$");
-        private HashSet<String> isShapeMatchedAlready;
+        private readonly Regex _shapeNamePattern = new Regex(@"^[^\[]\D+\s\d+$");
+        private HashSet<String> _isShapeMatchedAlready;
 
         private void AfterPasteEventHandler(PowerPoint.Selection selection)
         {
@@ -941,7 +950,7 @@ namespace PowerPointLabs
                         corruptedShapes[i].Delete();
                     }
 
-                    isShapeMatchedAlready = new HashSet<string>();
+                    _isShapeMatchedAlready = new HashSet<string>();
                     for (int i = 1; i <= pastedShapes.Count; i++)
                     {
                         PowerPoint.Shape shape = pastedShapes[i];
@@ -975,9 +984,9 @@ namespace PowerPointLabs
                     && IsSimilarName(shape.Name, _copiedShapes[i].Name)
                     && shape.Left == _copiedShapes[i].Left
                     && shape.Height == _copiedShapes[i].Height
-                    && !isShapeMatchedAlready.Contains(_copiedShapes[i].Id.ToString()))
+                    && !_isShapeMatchedAlready.Contains(_copiedShapes[i].Id.ToString()))
                 {
-                    isShapeMatchedAlready.Add(_copiedShapes[i].Id.ToString());
+                    _isShapeMatchedAlready.Add(_copiedShapes[i].Id.ToString());
                     return i;
                 }
             }
@@ -986,9 +995,9 @@ namespace PowerPointLabs
             {
                 if (IsSimilarShape(shape, _copiedShapes[i])
                     && IsSimilarName(shape.Name, _copiedShapes[i].Name)
-                    && !isShapeMatchedAlready.Contains(_copiedShapes[i].Id.ToString()))
+                    && !_isShapeMatchedAlready.Contains(_copiedShapes[i].Id.ToString()))
                 {
-                    isShapeMatchedAlready.Add(_copiedShapes[i].Id.ToString());
+                    _isShapeMatchedAlready.Add(_copiedShapes[i].Id.ToString());
                     return i;
                 }
             }
@@ -1157,12 +1166,8 @@ namespace PowerPointLabs
                 _eventHook = IntPtr.Zero;
             }
             if (eventType == (uint)Native.Event.EVENT_SYSTEM_MENUEND)
-            {
-                string description = "To activate 'Double Click to Open Property' feature, you need to enable 'Home' tab " +
-                              "in Options -> Customize Ribbon -> Main Tabs -> tick the checkbox of 'Home' -> click OK but" +
-                              "ton to save.";
-                string title = "Unable to activate 'Double Click to Open Property' feature";
-                MessageBox.Show(description, title);
+            {                
+                MessageBox.Show(TextCollection.TabActivateErrorDescription, TextCollection.TabActivateErrorTitle);
             }
         }
 
@@ -1172,6 +1177,7 @@ namespace PowerPointLabs
 
         private const string OfficeVersion2013 = "15.0";
         private const string OfficeVersion2010 = "14.0";
+        private const string ShortcutAltHO = "%ho";
 
         private const int CommandOpenBackgroundFormat = 0x8F;
 
@@ -1243,7 +1249,6 @@ namespace PowerPointLabs
             {
                 if (!_isInSlideShow)
                 {
-                    string Shortcut_Alt_H_O = "%ho";
                     if (_eventHook == IntPtr.Zero)
                     {
                         //Check whether Home tab is enabled or not
@@ -1256,7 +1261,7 @@ namespace PowerPointLabs
                             0,
                             0);
                     }
-                    SendKeys.Send(Shortcut_Alt_H_O);
+                    SendKeys.Send(ShortcutAltHO);
                 }
             }
             catch (InvalidOperationException)
